Implement updating the scores displayed
Display elements are referenced by frame & row. The reference is created
before recordScore is called on the scorecard object, using the frame/row
values appropriate to the score recorded, but the display is not updated
until after the recordScore call to allow for error trapping.

Also added a zero-score button for gutter balls.

jQuery removed as it is not currently needed.