Prevent adding further scores beyond the tenth frame
The test for this was specifically written without a strike in frame 10
as special consideration will be needed to handle this scenario when
strike bonuses are implemented.

General refactoring to reduce the size and responsibilities of certain
methods, and based on errors thrown by "use strict". Additionally,
updating the display during the tenth frame was causing problems as it
was trying to find a cell for the eleventh frame, due to the for loop
ending on scorecard.currentFrame which ticks over to 11 at the end of
the tenth frame; the simplest fix was to update all 10 frame scores on
every pass, as those which have not yet been played will now return and
display an empty string.