Allow scorecards to use bonusCounters for strikes
Scorecard tests are not currently isolated - consideration should be
given to rewriting them using spies, but this might involve replicating
full functionality inherited from BonusCounter since the objects persist
and are updated over the course of multiple frames.

Spares have not yet been given access to bonusCounters.

Note there is currently a hard block on rolling additional balls after
the tenth frame, this will need to be softened for strikes and spares.