Update README, fix bug
The bug relates to calculating the total score when extra balls are
rolled after the tenth frame (e.g. after a strike) - there was no check
to prevent scores from the '11th' or '12th' frame being added to the
total as well as being added to frames that take them as a bonus.