Display results of bonus rolls post-tenth frame
Bonus ball scores are shown conditionally - the fields where they appear
are blank by default. The logic is currently held in the Controller
script and relies on checking whether text has already been displayed
for the first bonus ball, as well as checking the frame and roll number
of the most recent recorded score to determine whether it should update
the empty fields; it may be worth recording the bonus rolls in a
different way to how other rolls are stored for ease of reference.