First 7 unit tests.
The model contains a rollHistory object for 11 frames of bowling. The 
model can also: register a roll, switch to a new frame when a frame is 
complete, not register an 11th frame if a strike or spare was made, and 
only register 1 roll in the 11th if a spare was made.