ADD: the frame function.
Will need to add the frame function to the scorecard, so that the player knows what
frame the game is on. Will also be needed to keep track of score given the player
may not get get 20 rolls if they get a strike and for use with the 10th frame which
complicates things later.