User stories: Enter roll score & calculate simple frame score
As a player,
so that I don't have to remember my score,
I want to enter the number of pins I knock down with a single roll into a scorecard.

As a player,
so that I don't have to calculate my score,
I want to see my current total score for a frame when I have made both rolls.