Get Jasmine working in browser as well as terminal
Looks like when you run Jasmine from the terminal it uses a bunch of
`require` and `module.exports` commands to load your code. If you run
code with these commands in the browser then it doesn't work - these are
Node commands.

So I made them safe for the browser by checking this, or similar for
`module.exports` before running those commands:
- `typeof(require) !== undefined`
I ended up having to use some global variables in my tests to make this
work (see lines 3 and 4 in `FeatureSpec.js` for example), which is not
ideal.

I also started making a `Frame` class here.