REFACTOR: Significant changes needed to make it work
Design issues meant that I need to refactor the code
will need to extract a roll and a frame function, not
currently working, but want to save progress in the
process to keep a log of where i am at the moment. TDD
will need to be reworked once the code is refactored.