UPDATE: Updated readme prior to push/pull request.
Not currently working, as tried to pull roll and frame out of the card, but that
led me down a dark hole, not sure how to link them all together. Have strayed
away from leading with tests, as clearly i didnt design it well enough initially
with diagrams. Would like to start investigating how to link the files (or if
that is necessary). Still a work in progress.