Calculate game status for 1 strike followed by 0s
Requires game to understand that a strike frame is complete even though
it only has one roll. So I created a `Frame` object that contains
logic to determine whether a frame is complete and to calculate the
frame score, and updated my `Game` object to use this to keep track of
frame statuses.