Refactor overloaded `Game.roll` method
In addition to its responsibility for adding rolls to frames, it was
also deciding whether it was necessary to start a new frame, so I
extracted that to its own private method.

Javascript method of getting the last element of an array is also not
pretty, so lets make a wrapper for that.