ADD: Implemented final frame management
Added feature tests for games with spares. Currently a spare in final game was problematic as it treated the final frame the same as others. Implemented a second isInPlay function for the final frame (isInPlayFinalFrame which is used by scorecard when the currentFrame is the final frame. Features test for spares is now passing.