ADD: User can not enter invalid rolls
added a validRoll function to detect invalid rolls. This is implemented in the roll function on scorecard at the beginning so invalid rolls do not change the state and will be ignored.
also tweaked the displayScore function in the interface to display final frame scores when available rather than waiting till the end of the frame.