Rearrange standalone jasmine folder structure
This was done so that the spec folder sits at the top level,
rather than buried several layers down. This will makes sense to
me as tests also need to be run by the version of Jasmin running
through node. I also moved the SpecRunner.html and src folders to
the root directory to make them more accessible and consistent
with where other project files are going to be.

As a result of this, the links in SpecRunner.html had to be
updated to reflect new file positions.