Remove node-based jasmine tests
Node modules are unrecognised by the standalone tests that run
with SpecRunner, so it was either-or. I've gone with the
standalone as the javascript is intended to run on the browser.

I found a guide to setting up grunt to run the SpecRunner tests
from the commandline here:-
alrw.github.io/integrating-jasmine-standalone-and-travis-ci-using-grunt/

This has now been set up so that travis can still run the tests.